requirements.txt
asgiref==3.11.0
Django==6.0.1
django-environ==0.12.0
psycopg==3.2.6
sqlparse==0.5.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-redis==5.4.0
gunicorn==23.0.0
